requirements.txt
google-generativeai>=0.8.0
transformers>=4.35.0
torch>=2.0.0
openai-whisper
webrtcvad>=2.0.10
pyttsx3>=2.90
numpy>=1.24.0
huggingface-hub>=0.19.0
av>=8.0.0
sounddevice
rich
datasets
speechbrain
librosa
